+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+- Moved LSP-server compilation and analysis to its own worker thread (1876)
+
 ### Deprecated
 ### Removed
 ### Fixed